fix(client): make await_ready resilient to heavy-load daemon startup
A production report showed `uffs --daemon start` failing with "Daemon did
not become ready in time / request timed out" on a 7-drive / 25M-record
machine whose load legitimately took 2m27s - 27s past the (fixed) 2-minute
client timeout - even though the daemon was healthy and fully loaded
moments later per `--daemon status`.

await_ready's `timeout` is now an idle budget instead of a hard wall-clock
cutoff: every DaemonStatus::Loading response whose drives_loaded advances
resets the deadline, so a daemon under heavy system load that keeps
visibly making progress is never killed by an arbitrary fixed cutoff -
only one that stalls for a full timeout window is. A hard 5x outer
ceiling still bounds the total wait. Applied to both the sync client (hit
by the CLI) and its async sibling, with new tests proving both the
extend-on-progress and still-times-out-on-stall behavior.

connect.rs split connect/auto-start/retry into connect_autostart.rs to
stay under the workspace's 800-LOC file policy.
